requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
python-dotenv
apscheduler
sqlalchemy
pydantic
